refactor: 패키지 변경, 스파크 애플리케이션과 dao 삭제
