--- src/main/java/chess/WebApplication.java
@@ -1,10 +1,0 @@
-package chess;
-
-import chess.web.ChessController;
-
-public class WebApplication {
-    public static void main(String[] args) {
-        ChessController chessController = new ChessController();
-        chessController.run();
-    }
-}
