+++ src/main/java/chess/console/ConsoleApplication.java
@@ -1,9 +1,6 @@
-package chess;
-
-import chess.domain.GameMachine;
+package chess.console;
 
 public class ConsoleApplication {
-
     public static void main(String[] args) {
         new GameMachine().run();
     }
