+++ src/main/java/chess/repository/PieceRepositoryImpl.java
@@ -1,4 +1,4 @@
-package chess.dao;
+package chess.repository;
 
 import chess.domain.piece.Piece;
 import chess.domain.position.Position;
@@ -13,7 +13,9 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSourceUtils;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public class PieceRepositoryImpl implements PieceRepository {
 
     private static final String TABLE_NAME = "piece";
