+++ src/main/java/chess/service/RoomService.java
@@ -1,6 +1,6 @@
 package chess.service;
 
-import chess.dao.RoomRepository;
+import chess.repository.RoomRepository;
 import chess.web.dto.RoomDto;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
