+++ src/test/java/chess/repository/BoardRepositoryImplTest.java
@@ -1,4 +1,4 @@
-package chess.dao;
+package chess.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
