+++ src/test/java/chess/repository/RoomRepositoryImplTest.java
@@ -1,4 +1,4 @@
-package chess.dao;
+package chess.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
